LEL-018 refactor(database): move `JournalCategoryEntity` to journal package
